+++ crates/paralegal-flow/src/ana/inline/graph.rs
@@ -1,14 +1,13 @@
 use crate::{
-    ana::algebra::{MirEquation, TypedLocal},
-    ir::{regal, GlobalLocation},
-    mir, serde, ty,
-    utils::{time, write_sep, DisplayViaDebug, FnResolution, TinyBitSet, TyCtxtExt},
+    ana::algebra::MirEquation,
+    ir::{regal, GlobalLocal, GlobalLocation},
+    mir, serde,
+    utils::{time, write_sep, DisplayViaDebug, FnResolution, TinyBitSet},
     Either, HashMap, HashSet, Location, TyCtxt,
 };
 
 use super::algebra;
 
-use paralegal_spdg::global_location::GlobalLocationS;
 use petgraph::prelude as pg;
 
 pub type ArgNum = u32;
@@ -175,127 +174,6 @@ impl std::fmt::Display for Edge {
                 Either::Right(()) => write!(f, "ctrl"),
             },
         )
-    }
-}
-
-/// A [`mir::Local`] but also tracks the precise call chain it is reachable
-/// from.
-#[derive(Clone, Copy, Debug)]
-pub struct GlobalLocal<'tcx> {
-    pub(super) local: mir::Local,
-    location: Option<GlobalLocation>,
-    pub ty: ty::Ty<'tcx>,
-}
-
-impl<'tcx> std::cmp::PartialEq for GlobalLocal<'tcx> {
-    fn eq(&self, other: &Self) -> bool {
-        self.location == other.location && self.local == other.local
-    }
-}
-
-impl<'tcx> std::cmp::PartialOrd for GlobalLocal<'tcx> {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(
-            self.location
-                .partial_cmp(&other.location())?
-                .then(self.local.partial_cmp(&other.local)?),
-        )
-    }
-}
-
-impl<'tcx> std::cmp::Eq for GlobalLocal<'tcx> {}
-
-impl<'tcx> std::cmp::Ord for GlobalLocal<'tcx> {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.location
-            .cmp(&other.location)
-            .then(self.local.cmp(&other.local))
-    }
-}
-
-impl<'tcx> std::hash::Hash for GlobalLocal<'tcx> {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.local.hash(state);
-        self.location.hash(state);
-    }
-}
-
-impl<'tcx> GlobalLocal<'tcx> {
-    /// Construct a new global local in a root function (no call chain)
-    pub fn at_root(tcx: TyCtxt<'tcx>, local: mir::Local, context: FnResolution<'tcx>) -> Self {
-        Self {
-            local,
-            location: None,
-            ty: context.local_ty(local.into(), tcx),
-        }
-    }
-
-    /// Construct a new global local relative to this call chain.
-    pub fn relative(
-        tcx: TyCtxt<'tcx>,
-        local: mir::Local,
-        location: GlobalLocation,
-        context: FnResolution<'tcx>,
-    ) -> Self {
-        Self {
-            local,
-            location: Some(location),
-            ty: context.local_ty(local.into(), tcx),
-        }
-    }
-
-    pub fn from_typed_local(
-        tcx: TyCtxt<'tcx>,
-        local: TypedLocal<'tcx>,
-        context: FnResolution<'tcx>,
-    ) -> Self {
-        Self {
-            local: local.local,
-            location: None,
-            ty: context.best_effort_normalize(tcx, local.ty()),
-        }
-    }
-
-    /// Guarantees that `result.location().is_some()`
-    pub fn add_location_frame(self, frame: GlobalLocationS) -> Self {
-        let Self {
-            local,
-            location,
-            ty,
-        } = self;
-        let location = location.map_or_else(
-            || GlobalLocation::single(frame.location, frame.function),
-            |prior| frame.relativize(prior),
-        );
-        Self {
-            ty,
-            local,
-            location: Some(location),
-        }
-    }
-
-    /// Access to the variable name.
-    #[inline]
-    pub fn local(self) -> mir::Local {
-        self.local
-    }
-
-    /// Access to the call chain.
-    #[inline]
-    pub fn location(self) -> Option<GlobalLocation> {
-        self.location
-    }
-}
-
-impl<'tcx> std::fmt::Display for GlobalLocal<'tcx> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:?} @ ", self.local)?;
-        if let Some(loc) = self.location {
-            write!(f, "{}", loc)
-        } else {
-            f.write_str("root")
-        }?;
-        write!(f, ": {:?}", self.ty)
     }
 }
