+++ crates/paralegal-flow/src/ir/mod.rs
@@ -3,4 +3,6 @@
 pub use paralegal_spdg::global_location::*;
 pub mod flows;
 pub use flows::*;
+pub mod local;
 pub mod regal;
+pub use local::*;
